+++ src-programs/parts.cpp
@@ -228,8 +228,8 @@ void assignPartNumbers(ScorePageSet& infiles, int page, Options& opts) {
    // at 1 rather than 0) for staff items on the page.
    int i, j, k, ksize;
 
-   for (i=0; i<sysmap.size(); i++) {
-      for (j=0; j<sysmap[i].size(); j++) {
+   for (i=0; i<(int)sysmap.size(); i++) {
+      for (j=0; j<(int)sysmap[i].size(); j++) {
          ksize = staffItems[sysmap[i][j]].size();
          for (k=0; k<ksize; k++) {
             if (staffItems[sysmap[i][j]][k]->getP9() != 0) {
